Python Project: Hunting Vampires
Please refer to 'contact.py' for my Python project. The document 'vampire_hunting.pdf' provides the assignment guidelines, while the remaining files are datasets and imported functions.

In summary, this project focuses on developing functions and using logical deduction to determine the final outcome based on several 'rules'. It emphasizes coding logic, including writing loops, functions, and conditions.

The project received a score of 97.5 out of 100.